Tests/LibWeb/Text/expected/input-range-drag-fractional-value.txt

click at x=10: value=0
click at x=25: value=0
click at x=33: value=0
click at x=50: value=1
click at x=75: value=1
